+++ examples/humanoid_step.py
@@ -24,7 +24,7 @@ from ltv_mpc import MPCProblem, solve_mpc
 
 
 @dataclass
-class HumanoidSteppingProblem:
+class Parameters:
 
     com_height: float = 0.8
     dsp_duration: float = 0.1  # [s]
@@ -37,23 +37,23 @@ class HumanoidSteppingProblem:
     start_pos: float = 0.0  # [m]
 
 
-def build_mpc_problem(problem: HumanoidSteppingProblem):
-    T = problem.horizon_duration / problem.nb_timesteps
-    nb_init_dsp_steps = int(round(problem.dsp_duration / T))
-    nb_init_ssp_steps = int(round(problem.ssp_duration / T))
-    nb_dsp_steps = int(round(problem.dsp_duration / T))
+def build_mpc_problem(params: Parameters):
+    T = params.horizon_duration / params.nb_timesteps
+    nb_init_dsp_steps = int(round(params.dsp_duration / T))
+    nb_init_ssp_steps = int(round(params.ssp_duration / T))
+    nb_dsp_steps = int(round(params.dsp_duration / T))
     state_matrix = np.array(
         [[1.0, T, T ** 2 / 2.0], [0.0, 1.0, T], [0.0, 0.0, 1.0]]
     )
     input_matrix = np.array([T ** 3 / 6.0, T ** 2 / 2.0, T])
     input_matrix = input_matrix.reshape((3, 1))
-    eta = problem.com_height / problem.gravity
+    eta = params.com_height / params.gravity
     zmp_from_state = np.array([1.0, 0.0, -eta])
     ineq_matrix = np.array([+zmp_from_state, -zmp_from_state])
-    cur_max = problem.start_pos + 0.5 * problem.foot_length
-    cur_min = problem.start_pos - 0.5 * problem.foot_length
-    next_max = problem.end_pos + 0.5 * problem.foot_length
-    next_min = problem.end_pos - 0.5 * problem.foot_length
+    cur_max = params.start_pos + 0.5 * params.foot_length
+    cur_min = params.start_pos - 0.5 * params.foot_length
+    next_max = params.end_pos + 0.5 * params.foot_length
+    next_min = params.end_pos - 0.5 * params.foot_length
     ineq_vector = [
         np.array([+1000.0, +1000.0])
         if i < nb_init_dsp_steps
@@ -62,7 +62,7 @@ def build_mpc_problem(problem: HumanoidSteppingProblem):
         else np.array([+1000.0, +1000.0])
         if i - nb_init_dsp_steps - nb_init_ssp_steps < nb_dsp_steps
         else np.array([+next_max, -next_min])
-        for i in range(problem.nb_timesteps)
+        for i in range(params.nb_timesteps)
     ]
     return MPCProblem(
         transition_state_matrix=state_matrix,
@@ -70,24 +70,28 @@ def build_mpc_problem(problem: HumanoidSteppingProblem):
         ineq_state_matrix=ineq_matrix,
         ineq_input_matrix=None,
         ineq_vector=ineq_vector,
-        initial_state=np.array([problem.start_pos, 0.0, 0.0]),
-        goal_state=np.array([problem.end_pos, 0.0, 0.0]),
-        nb_timesteps=problem.nb_timesteps,
+        initial_state=np.array([params.start_pos, 0.0, 0.0]),
+        goal_state=np.array([params.end_pos, 0.0, 0.0]),
+        nb_timesteps=params.nb_timesteps,
         terminal_cost_weight=1.0,
         stage_state_cost_weight=None,
         stage_input_cost_weight=1e-3,
     )
 
 
-def plot_mpc_solution(problem, mpc, solution):
-    t = np.linspace(0.0, problem.horizon_duration, problem.nb_timesteps + 1)
-    X = solution.states
-    eta = problem.com_height / problem.gravity
+def plot_mpc_solution(params, mpc_problem, plan):
+    t = np.linspace(0.0, params.horizon_duration, params.nb_timesteps + 1)
+    X = plan.states
+    eta = params.com_height / params.gravity
     zmp_from_state = np.array([1.0, 0.0, -eta])
     zmp = X.dot(zmp_from_state)
     pos = X[:, 0]
-    zmp_min = [x[0] if abs(x[0]) < 10 else None for x in mpc.ineq_vector]
-    zmp_max = [-x[1] if abs(x[1]) < 10 else None for x in mpc.ineq_vector]
+    zmp_min = [
+        x[0] if abs(x[0]) < 10 else None for x in mpc_problem.ineq_vector
+    ]
+    zmp_max = [
+        -x[1] if abs(x[1]) < 10 else None for x in mpc_problem.ineq_vector
+    ]
     zmp_min.append(zmp_min[-1])
     zmp_max.append(zmp_max[-1])
     pylab.ion()
@@ -101,7 +105,7 @@ def plot_mpc_solution(problem, mpc, solution):
 
 
 if __name__ == "__main__":
-    stepping_problem = HumanoidSteppingProblem()
-    mpc_problem = build_mpc_problem(stepping_problem)
-    solution = solve_mpc(mpc_problem, solver="quadprog")
-    plot_mpc_solution(stepping_problem, mpc_problem, solution)
+    params = Parameters()
+    mpc_problem = build_mpc_problem(params)
+    plan = solve_mpc(mpc_problem, solver="quadprog")
+    plot_mpc_solution(params, mpc_problem, plan)