+++ CHANGELOG.md
@@ -58,7 +58,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Initial import from [pymanoid](https://github.com/stephane-caron/pymanoid/blob/5158d8902df6265604cec5d790e96f0035575c7a/pymanoid/mpc.py)'s ``mpc.py``.
-- Humanoid stepping example
+- Humanoid step example
 - Triple integrator example
 
 ### Changed